done ft_strjoin, ft_substr, ft_strtrim and do ft_split

+++ ft_strjoin.c
@@ -2,9 +2,9 @@
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	unsigned int	i;
-	unsigned int	j;
-	char			*str;
+	size_t	i;
+	size_t	j;
+	char	*str;
 
 	i = ft_strlen(s1);
 	j = ft_strlen(s2);
